requirements.txt
alembic==1.4.3
aniso8601==6.0.0
appdirs==1.4.4
attrs==20.3.0
awscli==1.19.17
Babel==2.9.0
botocore==1.20.17
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
crypto==1.4.1
distlib==0.3.1
docutils==0.15.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
Naked==0.1.31
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.4
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.3
virtualenv==20.3.1
Werkzeug==1.0.1
WTForms==2.3.3
